Exclude files from code coverage, because they are static definitions and doesn't need to be covered. In future if they evolve to have logic then we can cover them.
