+++ Proximus/Workflow/Infrastructure/ProximusWorkflowState.cs
@@ -1,6 +1,9 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
+
 namespace Proximus
 {
+    [ExcludeFromCodeCoverage]
     internal class ProximusWorkflowState : WorkflowState
     {
         internal ProximusWorkflowState(Logger logger, WorkflowDatastore store) : base(logger, store)
