+++ Proximus/Workflow/Steps/GeoHashToDistance/Bounds.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Diagnostics.CodeAnalysis;
+
+namespace Proximus
+{
+    [ExcludeFromCodeCoverage]
+    public class Bound
+    {
+        public Location SW { get; set; }
+        public Location NE { get; set; }
+    }
+}
